Trait n_body::particle::RandomParticleGen

pub trait RandomParticleGen: Sync + Debug {
    fn gen(&self, rng: &mut ThreadRng) -> Particle;
    fn get_enum(&self) -> DistributionMethod;

    fn gen_multi(&self, count: u16) -> Vec<Particle> { ... }
    fn write(&self, path: PathBuf) -> Result<()> { ... }
}

a trait to define what a random particle generator is

Required Methods

fn gen(&self, rng: &mut ThreadRng) -> Particle

generate a random particle with given settings

fn get_enum(&self) -> DistributionMethod

convert to DistributionMethod

Provided Methods

fn gen_multi(&self, count: u16) -> Vec<Particle>

generate many Particles with the given settings this function is automatically written by rust for anything structure that implements this trait

fn write(&self, path: PathBuf) -> Result<()>

deserializes a RandomParticleGen from the specified .ron file