Function n_body::start_screen::start_screen

pub async fn start_screen(
    settings_path: PathBuf,
    method_path: PathBuf
) -> (Settings, Box<dyn RandomParticleGen>)

display the splash screen then display the configuration screen async means that this function can let the functions called after it start running even before this function is done executing. The other functions will need to wait if they get to a point where they need this functions return